requirements.txt
alembic==1.16.1
attr==0.3.2
ConfigParser==7.2.0
contextlib2==21.6.0
cryptography==45.0.3
docutils==0.21.2
Flask==3.1.1
Flask_Migrate==4.1.0
flask_sqlalchemy==3.1.1
flask_wtf==1.2.2
HTMLParser==0.0.2
importlib_metadata==8.0.0
injector==0.22.0
ipython==9.3.0
ipywidgets==8.1.7
Jinja2==3.1.6
jnius==1.1.0
keyring==25.6.0
langchain_community==0.3.24
langchain_core==0.3.63
langchain_openai==0.3.19
lockfile==0.12.2
mock==5.2.0
openai==1.84.0
Pillow==11.2.1
protobuf==6.31.1
pyOpenSSL==25.1.0
pytest==8.3.5
python-dotenv==1.1.0
railroad==0.5.0
Sphinx==8.2.3
thread==2.0.5
tornado==6.5.1
truststore==0.10.1
urllib3_secure_extra==0.1.0
WTForms==3.2.1
xmlrpclib==1.0.1
